Update ROTATE.ino
UHHHHHHHHHH I was working on this at work and then my supervisor came but basically I was adding a move forward function that would reset the amount it needs to move after it moves one foot and was gonna check to see if the camera was sending anything based on if it sent an empty array of [0,0,0,0] from the vision. It may be a broken state cuz I got interrupted.